feat: move menu components to stable
