+++ docs/src/layouts/MainLayout.astro
@@ -146,7 +146,6 @@ const githubEditUrl = `${CONFIG.GITHUB_EDIT_URL}/${currentFile}`;
       import 'media-chrome';
       import 'media-chrome/dist/media-theme-element';
       import 'media-chrome/dist/experimental/index';
-      import 'media-chrome/dist/experimental/menu/index';
     </script>
   </head>
   <body>
