+++ examples/vanilla/advanced.html
@@ -5,7 +5,6 @@
     <title>Media Chrome Advanced Video Usage Example</title>
     <script type="module" src="https://cdn.jsdelivr.net/npm/hls-video-element@1.1/+esm"></script>
     <script type="module" src="../../dist/index.js"></script>
-    <script type="module" src="../../dist/experimental/menu/index.js"></script>
     <style>
       /* Hide custom elements that are not defined yet */
       :not(:defined) {
