+++ examples/vanilla/control-elements/media-captions-menu.html
@@ -4,7 +4,6 @@
     <meta name="viewport" content="width=device-width">
     <title>Media Chrome Captions Menu</title>
     <script type="module" src="../../../dist/index.js"></script>
-    <script type="module" src="../../../dist/experimental/menu/index.js"></script>
     <style>
       .examples {
         margin-top: 20px;
