+++ examples/vanilla/control-elements/media-chrome-menu.html
@@ -3,8 +3,8 @@
   <head>
     <meta name="viewport" content="width=device-width" />
     <title>Media Chrome Menu</title>
-    <script type="module" src="../../../dist/experimental/menu/media-chrome-menu.js"></script>
-    <script type="module" src="../../../dist/experimental/menu/media-chrome-menu-item.js"></script>
+    <script type="module" src="../../../dist/media-chrome-menu.js"></script>
+    <script type="module" src="../../../dist/media-chrome-menu-item.js"></script>
     <style>
       [disabled] {
         opacity: 60%;
