+++ examples/vanilla/control-elements/media-playback-rate-menu.html
@@ -4,7 +4,6 @@
     <meta name="viewport" content="width=device-width">
     <title>Media Chrome Playback Rate Menu</title>
     <script type="module" src="../../../dist/index.js"></script>
-    <script type="module" src="../../../dist/experimental/menu/index.js"></script>
     <style>
       /* Hide custom elements that are not defined yet */
       :not(:defined) {
