+++ examples/vanilla/control-elements/media-rendition-menu.html
@@ -5,7 +5,6 @@
     <title>Media Chrome Rendition Menu</title>
     <script type="module" src="https://cdn.jsdelivr.net/npm/@mux/mux-video@canary/+esm"></script>
     <script type="module" src="../../../dist/index.js"></script>
-    <script type="module" src="../../../dist/experimental/menu/index.js"></script>
     <style>
       /** add styles to prevent CLS (Cumulative Layout Shift) */
       media-controller:not([audio]) {
