+++ examples/vanilla/media-elements/mux-video.html
@@ -5,26 +5,23 @@
     <title>Media Chrome &lt;mux-video&gt; Example</title>
     <script type="module" src="https://cdn.jsdelivr.net/npm/@mux/mux-video/+esm"></script>
     <script type="module" src="../../../dist/index.js"></script>
-    <script
-      type="module"
-      src="../../../../dist/experimental/menu/index.js"
-    ></script>
     <style>
       /** add styles to prevent CLS (Cumulative Layout Shift) */
       media-controller:not([audio]) {
         display: grid;          /* expands the container if preload=none */
         max-width: 800px;       /* allows the container to shrink if small */
         aspect-ratio: 16 / 9;   /* set container aspect ratio if preload=none */
-      }
-
-      mux-video {
         width: calc(100vmax - (8px * 2));
       }
 
       @media (min-width: 768px) {
-        mux-video {
+        media-controller:not([audio]) {
           width: 50vw;
         }
+      }
+
+      mux-video {
+        width: 100%;
       }
 
       .examples {
