--- src/js/experimental/menu/index.js
@@ -1,14 +1,0 @@
-export { MediaChromeDialog } from './media-chrome-dialog.js';
-export { MediaChromeMenu } from './media-chrome-menu.js';
-export { MediaChromeMenuItem } from './media-chrome-menu-item.js';
-export { MediaSettingsMenu } from './media-settings-menu.js';
-export { MediaSettingsMenuItem } from './media-settings-menu-item.js';
-export { MediaSettingsMenuButton } from './media-settings-menu-button.js';
-export { MediaAudioTrackMenu } from './media-audio-track-menu.js';
-export { MediaAudioTrackMenuButton } from './media-audio-track-menu-button.js';
-export { MediaCaptionsMenu } from './media-captions-menu.js';
-export { MediaCaptionsMenuButton } from './media-captions-menu-button.js';
-export { MediaPlaybackRateMenu } from './media-playback-rate-menu.js';
-export { MediaPlaybackRateMenuButton } from './media-playback-rate-menu-button.js';
-export { MediaRenditionMenu } from './media-rendition-menu.js';
-export { MediaRenditionMenuButton } from './media-rendition-menu-button.js';
