+++ src/js/index.js
@@ -2,55 +2,83 @@ export * as constants from './constants.js';
 export { default as labels } from './labels/labels.js';
 export * as timeUtils from './utils/time.js';
 import MediaAirplayButton from './media-airplay-button.js';
+import MediaAudioTrackMenu from './media-audio-track-menu.js';
+import MediaAudioTrackMenuButton from './media-audio-track-menu-button.js';
+import MediaCaptionsButton from './media-captions-button.js';
+import MediaCaptionsMenu from './media-captions-menu.js';
+import MediaCaptionsMenuButton from './media-captions-menu-button.js';
 import MediaCastButton from './media-cast-button.js';
 import MediaChromeButton from './media-chrome-button.js';
-import MediaGestureReceiver from './media-gesture-receiver.js';
-import MediaController from './media-controller.js';
+import MediaChromeDialog from './media-chrome-dialog.js';
+import MediaChromeMenu from './media-chrome-menu.js';
+import MediaChromeMenuItem from './media-chrome-menu-item.js';
 import MediaChromeRange from './media-chrome-range.js';
 import MediaControlBar from './media-control-bar.js';
+import MediaController from './media-controller.js';
 import MediaDurationDisplay from './media-duration-display.js';
-import MediaTimeDisplay from './media-time-display.js';
-import MediaCaptionsButton from './media-captions-button.js';
-import MediaSeekForwardButton from './media-seek-forward-button.js';
 import MediaFullscreenButton from './media-fullscreen-button.js';
+import MediaGestureReceiver from './media-gesture-receiver.js';
 import MediaLiveButton from './media-live-button.js';
+import MediaLoadingIndicator from './media-loading-indicator.js';
 import MediaMuteButton from './media-mute-button.js';
 import MediaPipButton from './media-pip-button.js';
-import MediaPlayButton from './media-play-button.js';
 import MediaPlaybackRateButton from './media-playback-rate-button.js';
+import MediaPlaybackRateMenu from './media-playback-rate-menu.js';
+import MediaPlaybackRateMenuButton from './media-playback-rate-menu-button.js';
+import MediaPlayButton from './media-play-button.js';
 import MediaPosterImage from './media-poster-image.js';
-import MediaSeekBackwardButton from './media-seek-backward-button.js';
-import MediaPreviewTimeDisplay from './media-preview-time-display.js';
-import MediaPreviewThumbnail from './media-preview-thumbnail.js';
 import MediaPreviewChapterDisplay from './media-preview-chapter-display.js';
+import MediaPreviewThumbnail from './media-preview-thumbnail.js';
+import MediaPreviewTimeDisplay from './media-preview-time-display.js';
+import MediaRenditionMenu from './media-rendition-menu.js';
+import MediaRenditionMenuButton from './media-rendition-menu-button.js';
+import MediaSeekBackwardButton from './media-seek-backward-button.js';
+import MediaSeekForwardButton from './media-seek-forward-button.js';
+import MediaSettingsMenu from './media-settings-menu.js';
+import MediaSettingsMenuButton from './media-settings-menu-button.js';
+import MediaSettingsMenuItem from './media-settings-menu-item.js';
+import MediaTimeDisplay from './media-time-display.js';
 import MediaTimeRange from './media-time-range.js';
-import MediaLoadingIndicator from './media-loading-indicator.js';
 import MediaVolumeRange from './media-volume-range.js';
 
 export {
   MediaAirplayButton,
+  MediaAudioTrackMenu,
+  MediaAudioTrackMenuButton,
+  MediaCaptionsButton,
+  MediaCaptionsMenu,
+  MediaCaptionsMenuButton,
   MediaCastButton,
   MediaChromeButton,
-  MediaGestureReceiver,
-  MediaController,
+  MediaChromeDialog,
+  MediaChromeMenu,
+  MediaChromeMenuItem,
   MediaChromeRange,
   MediaControlBar,
+  MediaController,
   MediaDurationDisplay,
-  MediaTimeDisplay,
-  MediaCaptionsButton,
-  MediaSeekForwardButton,
   MediaFullscreenButton,
+  MediaGestureReceiver,
   MediaLiveButton,
+  MediaLoadingIndicator,
   MediaMuteButton,
   MediaPipButton,
-  MediaPlayButton,
   MediaPlaybackRateButton,
+  MediaPlaybackRateMenu,
+  MediaPlaybackRateMenuButton,
+  MediaPlayButton,
   MediaPosterImage,
-  MediaSeekBackwardButton,
-  MediaPreviewTimeDisplay,
-  MediaPreviewThumbnail,
   MediaPreviewChapterDisplay,
+  MediaPreviewThumbnail,
+  MediaPreviewTimeDisplay,
+  MediaRenditionMenu,
+  MediaRenditionMenuButton,
+  MediaSeekBackwardButton,
+  MediaSeekForwardButton,
+  MediaSettingsMenu,
+  MediaSettingsMenuButton,
+  MediaSettingsMenuItem,
+  MediaTimeDisplay,
   MediaTimeRange,
-  MediaLoadingIndicator,
   MediaVolumeRange,
 };
