+++ src/js/media-audio-track-menu.js
@@ -1,6 +1,6 @@
-import { globalThis } from '../../utils/server-safe-globals.js';
-import { MediaUIAttributes, MediaUIEvents } from '../../constants.js';
-import { parseAudioTrackList } from '../../utils/utils.js';
+import { globalThis } from './utils/server-safe-globals.js';
+import { MediaUIAttributes, MediaUIEvents } from './constants.js';
+import { parseAudioTrackList } from './utils/utils.js';
 import {
   MediaChromeMenu,
   createMenuItem,
@@ -10,7 +10,7 @@ import {
   getStringAttr,
   setStringAttr,
   getMediaController,
-} from '../../utils/element-utils.js';
+} from './utils/element-utils.js';
 
 /**
  * @attr {string} mediaaudiotrackenabled - (read-only) Set to the enabled audio track.
