+++ src/js/media-captions-menu.js
@@ -1,6 +1,6 @@
-import { globalThis, document } from '../../utils/server-safe-globals.js';
-import { MediaUIAttributes, MediaUIEvents } from '../../constants.js';
-import { getMediaController } from '../../utils/element-utils.js';
+import { globalThis, document } from './utils/server-safe-globals.js';
+import { MediaUIAttributes, MediaUIEvents } from './constants.js';
+import { getMediaController } from './utils/element-utils.js';
 import {
   MediaChromeMenu,
   createMenuItem,
@@ -10,7 +10,7 @@ import {
   parseTextTracksStr,
   stringifyTextTrackList,
   formatTextTrackObj,
-} from '../../utils/captions.js';
+} from './utils/captions.js';
 
 const ccIcon = /*html*/`
   <svg aria-hidden="true" viewBox="0 0 26 24" part="captions-indicator indicator">
