+++ src/js/media-chrome-menu-button.js
@@ -1,7 +1,7 @@
-import { MediaChromeButton } from '../../media-chrome-button.js';
-import { globalThis } from '../../utils/server-safe-globals.js';
-import { InvokeEvent } from '../../utils/events.js';
-import { getDocumentOrShadowRoot } from '../../utils/element-utils.js';
+import { MediaChromeButton } from './media-chrome-button.js';
+import { globalThis } from './utils/server-safe-globals.js';
+import { InvokeEvent } from './utils/events.js';
+import { getDocumentOrShadowRoot } from './utils/element-utils.js';
 
 /**
  * @attr {string} invoketarget - The id of the element to invoke when clicked.
