+++ src/js/media-rendition-menu.js
@@ -1,7 +1,7 @@
-import { globalThis } from '../../utils/server-safe-globals.js';
-import { MediaUIAttributes, MediaUIEvents } from '../../constants.js';
-import { getMediaController, getStringAttr, setStringAttr } from '../../utils/element-utils.js';
-import { parseRenditionList } from '../../utils/utils.js';
+import { globalThis } from './utils/server-safe-globals.js';
+import { MediaUIAttributes, MediaUIEvents } from './constants.js';
+import { getMediaController, getStringAttr, setStringAttr } from './utils/element-utils.js';
+import { parseRenditionList } from './utils/utils.js';
 import {
   MediaChromeMenu,
   createMenuItem,
