+++ src/js/media-settings-menu-item.js
@@ -1,4 +1,4 @@
-import { globalThis, document } from '../../utils/server-safe-globals.js';
+import { globalThis, document } from './utils/server-safe-globals.js';
 import { MediaChromeMenuItem } from './media-chrome-menu-item.js';
 
 const template = document.createElement('template');
@@ -14,11 +14,13 @@ template.innerHTML = MediaChromeMenuItem.template.innerHTML + /*html*/`
   </style>
 `;
 
-template.content.querySelector('slot[name="suffix"]').innerHTML = /*html*/`
-  <svg aria-hidden="true" viewBox="0 0 20 24">
-    <path d="m8.12 17.585-.742-.669 4.2-4.665-4.2-4.666.743-.669 4.803 5.335-4.803 5.334Z"/>
-  </svg>
-`;
+if (template.content?.querySelector) {
+  template.content.querySelector('slot[name="suffix"]').innerHTML = /*html*/`
+    <svg aria-hidden="true" viewBox="0 0 20 24">
+      <path d="m8.12 17.585-.742-.669 4.2-4.665-4.2-4.666.743-.669 4.803 5.335-4.803 5.334Z"/>
+    </svg>
+  `;
+}
 
 class MediaSettingsMenuItem extends MediaChromeMenuItem {
   static template = template;
