+++ src/js/media-settings-menu.js
@@ -1,6 +1,6 @@
-import { globalThis, document } from '../../utils/server-safe-globals.js';
+import { globalThis, document } from './utils/server-safe-globals.js';
 import { MediaChromeMenu } from './media-chrome-menu.js';
-import { getMediaController } from '../../utils/element-utils.js';
+import { getMediaController } from './utils/element-utils.js';
 
 const template = document.createElement('template');
 // prettier-ignore
